pilot/NetworkSentinel/net_sentinel/requirements.txt
# requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
requests>=2.28.0
cryptography>=41.0.0
torch>=2.0.0  # For LLM operations
transformers>=4.30.0  # For LLM operations
pyyaml>=6.0
jinja2>=3.1.2  # For report generation
lxml>=4.9.0   # For XML processing
typing-extensions>=4.5.0
rich>=13.0.0  # For CLI formatting
python-dotenv>=1.0.0

# setup.py
from setuptools import setup, find_packages

setup(
    name="net-sentinel",
    version="1.0.0",
    packages=find_packages(),
    include_package_data=True,
    install_requires=[
        line.strip()
        for line in open("requirements.txt")
        if line.strip() and not line.startswith("#")
    ],
    entry_points={
        "console_scripts": [
            "net-sentinel=src.main:main",
        ],
    },
    python_requires=">=3.8",
    author="Your Name",
    author_email="your.email@example.com",
    description="Net-Sentinel - An Advanced Network Security Assessment Tool with AI Guidance",
    long_description=open("README.md").read(),
    long_description_content_type="text/markdown",
    keywords="security, penetration-testing, vulnerability-assessment, network-security",
    classifiers=[
        "Development Status :: 4 - Beta",
        "Intended Audience :: Information Technology",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Topic :: Security",
    ],
)